+++ src/main/java/stefan/licenta/homefinder/dto/NewAdDto.java
@@ -15,7 +15,7 @@ public class NewAdDto {
     private String adType;
     private String description;
     private Double price;
-    private Integer surface;
+    private Double surface;
     private Integer rooms;
     private String adItemType;
     private Double lat;
